fakevideo: add casting from double to uint64_t

+++ modules/fakevideo/fakevideo.c
@@ -51,7 +51,7 @@ static struct vidisp *vidisp;
 
 static void process_frame(struct vidsrc_st *st)
 {
-	st->ts += (VIDEO_TIMEBASE / st->fps);
+	st->ts += (uint64_t)(VIDEO_TIMEBASE / st->fps);
 
 	st->frameh(st->frame, st->ts, st->arg);
 }
